lang: use macro for Accounts trait impls
